pyproject.toml
[tool.poetry]
name = "deepl-tr"
version = "0.1.0a2"
description = "deepl translate using deepl-fastapi and webui2"
authors = ["ffreemt"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/ffreemt/deepl-tr-webui"

[tool.poetry.dependencies]
python = "^3.8.3"
logzero = "^1.7.0"
set-loglevel = "^0.1.2"
typer = "^0.4.1"
icecream = "^2.1.1"
install = "^1.3.5"
webui2 = "^2.0.6"
Jinja2 = "^3.1.2"
loguru = "^0.6.0"
cchardet = "^2.1.7"
charset-normalizer = "^3.0.1"
deepl-scraper-pp2 = "^0.1.0a2"
httpx = "^0.23.3"

[tool.poe.executor]
type = "poetry"

[tool.poe.tasks]
memo = "echo poe test or poetry run poe test"
test = "pytest tests"
pyright = "pyright deepl_tr"
flake8 = "flake8 deepl_tr --ignore F401,E501,F841"
check = ["pyright", "flake8"]
export = "poetry export --without-hashes -f requirements.txt -o requirements.txt"
build = "poetry build"
publish = "poetry publish"
release = ["test", "build", "publish"]
lint = { cmd = "pylint deepl_tr" }
isort = "isort tests deepl_tr"
black = "black tests deepl_tr"
format = ["isort", "black"]
docstyle = "pydocstyle --convention=google tests deepl_tr"
prerelease = {cmd = "poetry version prerelease && sync-version"}
tunnel = {cmd = "ssh -CN ip_or_hostname_defined_in_hosts -L 9091:127.0.0.1:9091"}

[tool.poetry.dev-dependencies]
pytest = "^7.1"
flake8 = "^3.9.2"
isort = "^5.10.1"
pylint = "^2.14.2"
pyright = "^1.1.254"
pydocstyle = "^6.1.1"
toml = "^0.10.2"
# tbump = "^6.3.2"
poethepoet = "^0.10.0"

[tool.poetry.scripts]
"deepl-tr" = "deepl_tr.__main__:app"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --doctest-modules"
log_cli = true
testpaths = [
    "tests",
    # "integration",
]

[tool.black]
skip-string-normalization = 0

[tool.isort]
profile = "black"
multi_line_output = 3
